Move HazelcastAccessor to test scope
