+++ hazelcast-hibernate5/src/test/java/com/hazelcast/hibernate/CustomPropertiesTest.java
@@ -18,7 +18,6 @@ package com.hazelcast.hibernate;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.Config;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.hibernate.instance.HazelcastAccessor;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.SlowTest;
 import org.hamcrest.BaseMatcher;
