+++ hazelcast-hibernate52/src/test/java/com/hazelcast/hibernate/LocalRegionFactoryDefaultTest.java
@@ -48,7 +48,7 @@ public class LocalRegionFactoryDefaultTest extends RegionFactoryDefaultTest {
 
     @Test
     public void testEntity() {
-        final HazelcastInstance hz = getHazelcastInstance(sf);
+        final HazelcastInstance hz = HazelcastAccessor.getHazelcastInstance(sf);
         assertNotNull(hz);
         final int count = 100;
         final int childCount = 3;
