+++ hazelcast-hibernate53/src/test/java/com/hazelcast/hibernate/HibernateTestSupport.java
@@ -16,11 +16,9 @@
 package com.hazelcast.hibernate;
 
 import com.hazelcast.core.Hazelcast;
-import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.hibernate.entity.AnnotatedEntity;
 import com.hazelcast.hibernate.entity.DummyEntity;
 import com.hazelcast.hibernate.entity.DummyProperty;
-import com.hazelcast.hibernate.instance.HazelcastAccessor;
 import com.hazelcast.hibernate.instance.HazelcastMockInstanceFactory;
 import com.hazelcast.hibernate.instance.IHazelcastInstanceLoader;
 import com.hazelcast.internal.nio.IOUtil;
@@ -99,10 +97,6 @@ public abstract class HibernateTestSupport extends HazelcastTestSupport {
         sf.getStatistics().setStatisticsEnabled(true);
 
         return sf;
-    }
-
-    protected HazelcastInstance getHazelcastInstance(final SessionFactory sf) {
-        return HazelcastAccessor.getHazelcastInstance(sf);
     }
 
     protected void deleteDummyEntity(SessionFactory sf, long id) throws Exception {
